Rework args check
